Refactor electronStartup into multiple files
